Re-org attestation program
