+++ contracts/solana/programs/attestation/src/consts.rs
@@ -1,7 +1,8 @@
-use anchor_lang::prelude::*;
+// use anchor_lang::prelude::*;
 
-#[constant]
-pub const ADMIN_PUBLIC_KEY: &str = "5hswvHkD3hUuUrLdfVRM59BmR46muY3Tf2J7UsyvZ5DL";
+// #[constant]
+// pub const ADMIN_PUBLIC_KEY: &str = "5hswvHkD3hUuUrLdfVRM59BmR46muY3Tf2J7UsyvZ5DL";
+
 // const ADMIN_PUBLIC_KEY: Pubkey = solana_program::pubkey!("5hswvHkD3hUuUrLdfVRM59BmR46muY3Tf2J7UsyvZ5DL")
 // Fetch the program ID dynamically from the environment or context
 // #[constant]
