+++ contracts/solana/programs/attestation/src/errors.rs
@@ -1,0 +1,17 @@
+use anchor_lang::prelude::*;
+
+#[error_code]
+pub enum AttestationError {
+    #[msg("Invalid Schema")]
+    InvalidSchema,
+    #[msg("Attestation not found.")]
+    NotFound,
+    #[msg("Attestation already revoked.")]
+    AlreadyRevoked,
+    #[msg("Schema is not revocable.")]
+    Irrevocable,
+    #[msg("Invalid expiration time.")]
+    InvalidExpirationTime,
+    #[msg("Data too large.")]
+    DataTooLarge,
+}
