+++ contracts/solana/programs/attestation/src/instructions.rs
@@ -1,0 +1,7 @@
+mod attest;
+mod initialize;
+mod revoke;
+
+pub use attest::*;
+pub use initialize::*;
+pub use revoke::*;
